Moves random stub to keystest package
