+++ core/internal/testutils/keystest/seedable_rand_reader.go
@@ -1,21 +1,20 @@
-package mocks
+package keystest
 
 import (
 	"io"
-	//nolint
-	mrand "math/rand"
+	"math/rand"
 )
 
 type randReader struct{}
 
 func (randReader) Read(b []byte) (n int, err error) {
-	return mrand.Read(b)
+	return rand.Read(b)
 }
 
 // NewRandReaderFromSeed returns a seedable random io reader, producing deterministic
 // output. This is useful for deterministically producing keys for tests. This is an
 // insecure source of randomness and therefor should only be used in tests.
 func NewRandReaderFromSeed(seed int64) io.Reader {
-	mrand.Seed(seed)
+	rand.Seed(seed)
 	return randReader{}
 }
